#1883 Move GOSymbolTable to a standalone file.
