+++ src/psyclone/domain/gocean/__init__.py
@@ -1,7 +1,7 @@
 # -----------------------------------------------------------------------------
 # BSD 3-Clause License
 #
-# Copyright (c) 2021, Science and Technology Facilities Council.
+# Copyright (c) 2021-2022, Science and Technology Facilities Council.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -37,9 +37,10 @@
 '''
 
 from psyclone.domain.gocean.gocean_constants import GOceanConstants
+from psyclone.domain.gocean.go_symbol_table import GOSymbolTable
 
 # The entities in the __all__ list are made available to import directly from
 # this package e.g.:
 # from psyclone.domain.gocean import GOceanConstants
 
-__all__ = ['GOceanConstants']
+__all__ = ['GOceanConstants', 'GOSymbolTable']
